Output the PNG the right way up
The output is calculated column-by-column but the PNG file is written
row-by-row, so this change stores all output as it is calculated and
regurgitates it all at the end. This also allows us to accumulate the
maximum output value to be able to notmalise the output so that the
loudest pixel gets white as a colour.